Created StartForScripts.sh
This is a file that is used when a modpack (or similar) gives a start script but you wish to run that script in a screen. This script serves as a bridge to add that screen functionality without meddling with anything else and allowing the given script to do what it wants to do.